deployment/requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
pillow
tensorflow